+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Improvements
 - PR #903 Add short commit hash to conda package
+- PR #920 modify bfs test, update graph number_of_edges, update storage of transposedAdjList in Graph
 
 ## Bug Fixes